Move investigation HOI selectors to a sub-directory
- Avoid confusion with screening selectors of the same name

[#152464349]
