+++ app/javascript/investigations/HistoryOfInvolvementContainer.jsx
@@ -1,6 +1,6 @@
 import {connect} from 'react-redux'
 import HistoryOfInvolvement from 'views/history/HistoryOfInvolvement'
-import {getHistoryIsEmptySelector} from 'selectors/historyOfInvolvementSelectors'
+import {getHistoryIsEmptySelector} from 'selectors/investigation/historyOfInvolvementSelectors'
 
 const mapStateToProps = (state) => (
   {historyIsEmpty: getHistoryIsEmptySelector(state)}
